[ADD] Added main screen with updating order line

+++ app/src/main/java/com/cronyapps/odoo/core/orm/type/FieldFloat.java
@@ -2,7 +2,7 @@ package com.cronyapps.odoo.core.orm.type;
 
 import com.cronyapps.odoo.core.orm.utils.FieldType;
 
-public class FieldFloat extends FieldType<FieldFloat, Double> {
+public class FieldFloat extends FieldType<FieldFloat, Float> {
 
     public FieldFloat(String label) {
         super(label);
